Remove TryIntoSubstates trait
