+++ radix-engine/src/model/mod.rs
@@ -1,6 +1,7 @@
 mod abi_extractor;
 mod auth_converter;
 mod method_authorization;
+mod node_to_substates;
 mod nodes;
 mod package_extractor;
 mod resource;
@@ -10,6 +11,7 @@ pub use crate::engine::InvokeError;
 pub use abi_extractor::*;
 pub use auth_converter::convert;
 pub use method_authorization::*;
+pub use node_to_substates::*;
 pub use nodes::*;
 pub use package_extractor::{extract_abi, ExtractAbiError};
 pub use resource::*;
