+++ radix-engine/src/model/nodes/component.rs
@@ -4,8 +4,6 @@ use crate::model::{ComponentInfoSubstate, ComponentStateSubstate, InvokeError};
 use crate::types::*;
 use crate::wasm::{WasmEngine, WasmInstance};
 
-use super::TryIntoSubstates;
-
 #[derive(Debug, TypeId, Encode, Decode)]
 pub enum ComponentError {
     InvalidRequestData(DecodeError),
@@ -19,17 +17,6 @@ pub struct Component {
     // We use the state to look up children nodes of a component.
     // Changing to lazily loading at this moment requires an even larger refactoring to the code base.
     pub state: ComponentStateSubstate,
-}
-
-impl TryIntoSubstates for Component {
-    type Error = ();
-
-    fn try_into_substates(self) -> Result<Vec<crate::model::Substate>, Self::Error> {
-        let mut substates = Vec::new();
-        substates.push(self.info.into());
-        substates.push(self.state.into());
-        Ok(substates)
-    }
 }
 
 impl Component {
