+++ radix-engine/src/model/nodes/key_value_store.rs
@@ -1,7 +1,5 @@
 use crate::types::*;
 
-use super::TryIntoSubstates;
-
 #[derive(Debug)]
 pub struct HeapKeyValueStore {
     pub store: HashMap<Vec<u8>, ScryptoValue>,
@@ -20,13 +18,5 @@ impl HeapKeyValueStore {
 
     pub fn get(&self, key: &[u8]) -> Option<ScryptoValue> {
         self.store.get(key).cloned()
-    }
-}
-
-impl TryIntoSubstates for HeapKeyValueStore {
-    type Error = ();
-
-    fn try_into_substates(self) -> Result<Vec<crate::model::Substate>, Self::Error> {
-        Ok(vec![])
     }
 }
