+++ radix-engine/src/model/nodes/mod.rs
@@ -23,9 +23,3 @@ pub use system::*;
 pub use transaction_processor::*;
 pub use vault::*;
 pub use worktop::*;
-
-pub trait TryIntoSubstates {
-    type Error;
-
-    fn try_into_substates(self) -> Result<Vec<crate::model::Substate>, Self::Error>;
-}
